crowd_monitoring_system/requirements.txt
ultralytics
opencv-python
pandas
prophet
fastapi
uvicorn
streamlit
plotly
streamlit-webrtc
python-multipart
requests
python-dotenv
pytest
